A Look at University Completion Rates
I will be focusing on the information on the average rate of student completion for four-year institutions. I will be checking for a relationship between this rate and the type of institution (public, private and non-profit, or private and for-profit) and how students pay for their education (whether they received a Pell grant, took out a federal loan, or did neither).